fix(Transformer.js): support webservers that provide correct MIME types
Transformer.js expects to get the BPMN XML as string, which works with
servers that deliver *.bpmn files as application/octet-stream, e.g.
Node.js, but fails on servers that correctly deliver them as text/xml.
In the latter case the BPMN XML will be already provided as a DOM
Document.