impr(executable): add new property "executable" to render options
This commit adds a new property called "executable" to the render options
to indicate whether the bpmn 2.0 xml should be executed by the
JavaScript-Process-Engine if it is set to "true".

In the other case (if the flag is set to "false") their are no further checks
whether the bpmn 2.0 xml can be executed by the JavaScript-Process-Engine.

The default value is "false"

related to CAM-1983